Answer to Marc's comments.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/gudhi/branches/gudhi_stat@2321 636b058d-ea47-450e-bf9e-a15bfbe3eedb

Former-commit-id: 4eb08a7ca5664fffd8e683c30871d3fd53d1a551

+++ src/Gudhi_stat/test/persistence_lanscapes_on_grid_test.cpp
@@ -56,8 +56,8 @@ BOOST_AUTO_TEST_CASE(check_construction_of_landscape)
 BOOST_AUTO_TEST_CASE(check_construction_of_landscape_using_only_ten_levels) 
 {	
 	//TODO
-	size_t number = 10;
-	Persistence_landscape_on_grid l( "data/file_with_diagram_1" , 100  ,number );	
+	unsigned number = 10;
+	Persistence_landscape_on_grid l( "data/file_with_diagram_1" , 100  ,number );		
 	Persistence_landscape_on_grid g( "data/file_with_diagram_1" , 100 );
 	//cut all the elements of order > 10 in g. 
 	
@@ -70,8 +70,8 @@ BOOST_AUTO_TEST_CASE(check_construction_of_landscape_using_only_ten_levels)
 		{
 			BOOST_CHECK( v1[i] == v2[i] );
 		}
-	}	
-}
+	}
+	}
